requirements.txt
app==0.0.1
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
cssutils==2.9.0
dkimpy==1.1.5
dnspython==2.6.1
Flask==3.0.2
idna==3.6
itsdangerous==2.1.2
jaraco.classes==3.3.1
Jinja2==3.1.3
jsonify==0.5
keyring==24.3.1
lxml==5.1.0
MarkupSafe==2.1.5
more-itertools==10.2.0
pillow==10.2.0
premailer==3.10.0
pymongo==4.6.2
pywin32-ctypes==0.2.2
requests==2.31.0
urllib3==2.2.1
Werkzeug==3.0.1
yagmail==0.15.293
